(cmdline) kernel-info: enable selinux but prevent enforcing
Droidian on vayu, as other devices, has issues when selinux is enabled.
The recommendation is disable it.

By other hand, some Droidian kernels has legacy apparmor patches
from Ubuntu Ports, and apparmor is configured as default security.
But apparmor seems issn't used by Droidian.

On this vayu port selinux will be used as default security and enabled
in the cmdline, but preventing enforcing.

The idea is ensure a reliable security infrastructure.

This prepare the kernel to implement full selinux support in the future.

Cmdline options: 'selinux=1 androidboot.selinux=permissive enforcing=0'

Signed-off-by: berbascum <berbascum@ticv.cat>